src/requirements-dev.txt
Django
djangorestframework
psycopg2
plotly
pdfkit
pandas
uuid